[IMP] convert to native javascript module
* convert PartnerDetailsEdit.js to native javascript module.
* remove usesell PartnerDetailsEdit.xml.
